refactor: reduce duplicated code in 3d renderer
